SCB-527 adjust spring boot structure
